+++ losttime.web/src/results2/CompetitionClass/SingleRaceSoloPointedResult.ts
