+++ losttime.web/src/results2/CompetitionClass/SingleRaceSoloResult.ts
