+++ losttime.web/src/results2/CompetitionClass/Templates/CompetitionClassPresetButton.ts
