+++ losttime.web/src/results2/Styles/PlaintextTable.ts
@@ -1,6 +1,6 @@
 import { PlaintextColumn } from "./PlaintextColumn";
-import { SingleRaceSoloPointedResult } from "./ResultTypes/SingleRaceSoloPointedResult";
-import { SingleRaceSoloResult } from "./ResultTypes/SingleRaceSoloResult";
+import { SingleRaceSoloPointedResult } from "../CompetitionClass/SingleRaceSoloPointedResult";
+import { SingleRaceSoloResult } from "../CompetitionClass/SingleRaceSoloResult";
 
 
 export class PlaintextTable {
