+++ losttime.web/src/results2/Styles/RenderStyles.ts
